django/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
Pillow==11.3.0
python-decouple==3.8
django-cors-headers==4.7.0
channels==4.2.0
channels-redis==4.2.0
redis==5.0.8
daphne==4.1.2
